Add copyWith and addCalendarDays to DateTimeBasics
As evidenced by a number of questions on StackOverflow, it's not
uncommon for people to want to perform date calculations with
logical, calendar days instead of with 24-hour increments.  They then
often get confused about why their code computes slightly unexpected
times at certain points of the year (and not noticing that those
points span DST changes).  Add an `addCalendarDays` extension method.

There also are a number of cases where copying a `DateTime` would be
useful, but doing so is tedious since a general function to copy a
`DateTime` needs separate code paths for UTC and non-UTC `DateTime`
objects.  Add a `copyWith` extension method, which also is useful for
implementing `addCalendarDays`.