Add an `addCalendarMonths` extension for `DateTime`
Add an `addCalendarMonths` extension method for `DateTime`. (It
probably could be named just `addMonths`, but I chose
`addCalendarMonths` to match the `addCalendarDays` extension
method.)

Like `addCalendarDays`, `addCalendarMonths` will preserve more
granular fields when possible.  We'll define the result of "adding N
months" to mean the normalized `DateTime` with N added to the month
value such that the day of the month either is preserved or is the
last day of that month, whichever is earlier. (See the examples from
the documentation.)